🌊 Feat: 가입 관련 뷰 열리면 처음 텍스트필드에 다 focus 되게 구현
**EnterPhoneNumberView**
- 전화번호 텍스트필드의 길이가 충족되지 않으면 버튼 활성화 안되게 수정
